# internal/api/docs/api_swagger.yaml
basePath: /api/
definitions:
  models.Task:
    properties:
      dateCreated:
        type: string
      description:
        type: string
      id:
        type: integer
      status:
        type: integer
      title:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    required:
    - title
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      status:
        type: integer
    type: object
  schemas.AuthConfirmation:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  schemas.AuthLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  schemas.AuthLoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  schemas.AuthRefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  schemas.AuthRefreshResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  schemas.AuthRegistrationRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  schemas.Error:
    properties:
      message: {}
      validationErrors: {}
    type: object
  schemas.TasksListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      total_count:
        type: integer
    type: object
host: localhost
info:
  contact: {}
  description: gotodo api
  title: gotodo
  version: "1.0"
paths:
  /auth/confirmation/:
    post:
      consumes:
      - application/json
      description: User confirmation after registration
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.AuthConfirmation'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Confirmation
      tags:
      - auth
  /auth/login/:
    post:
      consumes:
      - application/json
      description: Login in new sessions
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AuthLoginResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Login
      tags:
      - auth
  /auth/refresh/:
    post:
      consumes:
      - application/json
      description: Refresh exists session
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.AuthRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AuthRefreshResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Refresh
      tags:
      - auth
      - auth
  /auth/registration/:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.AuthRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Registration
      tags:
      - auth
  /tasks/:
    get:
      consumes:
      - application/json
      description: Get tasks list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.TasksListResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      security:
      - ApiHeaderAuth: []
      summary: List
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create new tasks
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      security:
      - ApiHeaderAuth: []
      summary: Create
      tags:
      - tasks
  /tasks/{id}/:
    delete:
      consumes:
      - application/json
      description: Delete task
      parameters:
      - description: Task identifier
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "404":
          description: ""
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      security:
      - ApiHeaderAuth: []
      summary: Delete
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get task details
      parameters:
      - description: Task identifier
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: ""
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      security:
      - ApiHeaderAuth: []
      summary: Retrieve
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update task
      parameters:
      - description: Task identifier
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/schemas.Error'
      security:
      - ApiHeaderAuth: []
      summary: Update
      tags:
      - tasks
securityDefinitions:
  ApiKeyHeader:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: Auth endpoints
  name: auth
- description: Tasks endpoint
  name: tasks